fix(mmkv): fix get bug by using default storage id

+++ src/features/auth/storages/AuthTokenStorage.ts
@@ -14,6 +14,8 @@ export class AuthTokenStorage {
   }
 
   public static delete(): void {
-    mmkvStorage.delete(authTokenStorageKey);
+    if (mmkvStorage.contains(authTokenStorageKey)) {
+      mmkvStorage.delete(authTokenStorageKey);
+    }
   }
 }
